+++ MSGraph/functions/exchange/Send-MgaMailMessage.ps1
@@ -11,6 +11,54 @@
         Carrier object for Pipeline input.
         This can be the id of the message or a message object passed in.
 
+    .PARAMETER Subject
+        The subject of the new message.
+
+    .PARAMETER Sender
+        The account that is actually used to generate the message.
+        (Updatable only when sending a message from a shared mailbox or sending a message as a delegate.
+        In any case, the value must correspond to the actual mailbox used.)
+
+    .PARAMETER From
+        The mailbox owner and sender of the message.
+        Must correspond to the actual mailbox used.
+
+    .PARAMETER ToRecipients
+        The To recipients for the message.
+
+    .PARAMETER CCRecipients
+        The Cc recipients for the message.
+
+    .PARAMETER BCCRecipients
+        The Bcc recipients for the message.
+
+    .PARAMETER ReplyTo
+        The email addresses to use when replying.
+
+    .PARAMETER Body
+        The body of the message.
+
+    .PARAMETER Categories
+        The categories associated with the message.
+
+    .PARAMETER Importance
+        The importance of the message.
+        The possible values are: Low, Normal, High.
+
+    .PARAMETER InferenceClassification
+        The classification of the message for the user, based on inferred relevance or importance, or on an explicit override.
+        The possible values are: focused or other.
+
+    .PARAMETER IsDeliveryReceiptRequested
+        Indicates whether a delivery receipt is requested for the message.
+
+    .PARAMETER IsReadReceiptRequested
+        Indicates whether a read receipt is requested for the message.
+
+    .PARAMETER SaveToSentItems
+        Indicates whether to save the message in Sent Items.
+        Only needed to be specified if the parameter should be $false, default is $true.
+
     .PARAMETER User
         The user-account to access. Defaults to the main user connected as.
         Can be any primary email name of any user the connected token has access to.
@@ -40,14 +88,89 @@
         PS C:\> $mail = New-MgaMailMessage -ToRecipients 'someone@something.org' -Subject 'A new Mail' -Body 'This is a new mail'
 
     #>
-    [CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'Medium')]
+    [CmdletBinding(SupportsShouldProcess = $true, ConfirmImpact = 'Medium', DefaultParameterSetName = 'DirectSend')]
     [Alias()]
     [OutputType([MSGraph.Exchange.Mail.Message])]
     param (
-        [Parameter(Mandatory=$true, Position = 0, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true)]
+        [Parameter(Mandatory = $true, Position = 0, ValueFromPipeline = $true, ValueFromPipelineByPropertyName = $true, ParameterSetName = 'ByInputObject')]
         [Alias('InputObject', 'MessageId', 'Id', 'Mail', 'MailId')]
         [MSGraph.Exchange.Mail.MessageParameter[]]
         $Message,
+
+        [Parameter(ParameterSetName = 'DirectSend', Mandatory = $true)]
+        [Alias('Name', 'Title')]
+        [string[]]
+        $Subject,
+
+        [Parameter(ParameterSetName = 'DirectSend')]
+        [AllowNull()]
+        [AllowEmptyCollection()]
+        [AllowEmptyString()]
+        [string]
+        $Sender,
+
+        [Parameter(ParameterSetName = 'DirectSend')]
+        [AllowNull()]
+        [AllowEmptyCollection()]
+        [AllowEmptyString()]
+        [string]
+        $From,
+
+        [Parameter(ParameterSetName = 'DirectSend', Mandatory = $true)]
+        [Alias('To', 'Recipients')]
+        [string[]]
+        $ToRecipients,
+
+        [Parameter(ParameterSetName = 'DirectSend')]
+        [AllowNull()]
+        [AllowEmptyCollection()]
+        [AllowEmptyString()]
+        [string[]]
+        $CCRecipients,
+
+        [Parameter(ParameterSetName = 'DirectSend')]
+        [AllowNull()]
+        [AllowEmptyCollection()]
+        [AllowEmptyString()]
+        [string[]]
+        $BCCRecipients,
+
+        [Parameter(ParameterSetName = 'DirectSend')]
+        [AllowNull()]
+        [AllowEmptyCollection()]
+        [AllowEmptyString()]
+        [string[]]
+        $ReplyTo,
+
+        [Parameter(ParameterSetName = 'DirectSend')]
+        [String]
+        $Body,
+
+        [Parameter(ParameterSetName = 'DirectSend')]
+        [String[]]
+        $Categories,
+
+        [Parameter(ParameterSetName = 'DirectSend')]
+        [ValidateSet("Low", "Normal", "High")]
+        [String]
+        $Importance,
+
+        [Parameter(ParameterSetName = 'DirectSend')]
+        [ValidateSet("focused", "other")]
+        [String]
+        $InferenceClassification,
+
+        [Parameter(ParameterSetName = 'DirectSend')]
+        [bool]
+        $IsDeliveryReceiptRequested,
+
+        [Parameter(ParameterSetName = 'DirectSend')]
+        [bool]
+        $IsReadReceiptRequested,
+
+        [Parameter(ParameterSetName = 'DirectSend')]
+        [bool]
+        $SaveToSentItems = $true,
 
         [string]
         $User,
@@ -64,32 +187,83 @@
     }
 
     process {
-        #Write-PSFMessage -Level Debug -Message "Working on parameter set $($PSCmdlet.ParameterSetName)" -Tag "ParameterSetHandling"
+        Write-PSFMessage -Level Debug -Message "Working on parameter set $($PSCmdlet.ParameterSetName)" -Tag "ParameterSetHandling"
 
         #region send message
-        foreach ($messageItem in $Message) {
-            #region checking input object type and query message if required
-            if ($messageItem.TypeName -like "System.String") {
-                $messageItem = Resolve-MailObjectFromString -Object $messageItem -User $User -Token $Token -NoNameResolving -FunctionName $MyInvocation.MyCommand
-                if(-not $messageItem) { continue }
-            }
+        switch ($PSCmdlet.ParameterSetName) {
+            'ByInputObject' {
+                foreach ($messageItem in $Message) {
+                    #region checking input object type and query message if required
+                    if ($messageItem.TypeName -like "System.String") {
+                        $messageItem = Resolve-MailObjectFromString -Object $messageItem -User $User -Token $Token -NoNameResolving -FunctionName $MyInvocation.MyCommand
+                        if (-not $messageItem) { continue }
+                    }
 
-            $User = Resolve-UserInMailObject -Object $messageItem -User $User -ShowWarning -FunctionName $MyInvocation.MyCommand
-            #endregion checking input object type and query message if required
+                    $User = Resolve-UserInMailObject -Object $messageItem -User $User -ShowWarning -FunctionName $MyInvocation.MyCommand
+                    #endregion checking input object type and query message if required
 
-            if ($pscmdlet.ShouldProcess($messageItem, "Send")) {
-                Write-PSFMessage -Tag "MessageSend" -Level Verbose -Message "Send message '$($messageItem)'"
-                $invokeParam = @{
-                    "Field"        = "messages/$($messageItem.Id)/send"
-                    "User"         = $User
-                    "Body"         = ""
-                    "ContentType"  = "application/json"
-                    "Token"        = $Token
-                    "FunctionName" = $MyInvocation.MyCommand
+                    if ($pscmdlet.ShouldProcess($messageItem, "Send")) {
+                        Write-PSFMessage -Tag "MessageSend" -Level Verbose -Message "Send message '$($messageItem)'"
+                        $invokeParam = @{
+                            "Field"        = "messages/$($messageItem.Id)/send"
+                            "User"         = $User
+                            "Body"         = ""
+                            "ContentType"  = "application/json"
+                            "Token"        = $Token
+                            "FunctionName" = $MyInvocation.MyCommand
+                        }
+                        $null = Invoke-MgaPostMethod @invokeParam
+                        if ($PassThru) { $messageItem.InputObject }
+                    }
                 }
-                $null = Invoke-MgaPostMethod @invokeParam
-                if ($PassThru) { $messageItem.InputObject }
             }
+
+            'DirectSend' {
+                $jsonParams = @{
+                    Subject                    = $Subject
+                    Sender                     = $Sender
+                    From                       = $From
+                    ToRecipients               = $ToRecipients
+                    CCRecipients               = $CCRecipients
+                    BCCRecipients              = $BCCRecipients
+                    ReplyTo                    = $ReplyTo
+                    Body                       = $Body
+                    Categories                 = $Categories
+                    Importance                 = $Importance
+                    InferenceClassification    = $InferenceClassification
+                    IsDeliveryReceiptRequested = $IsDeliveryReceiptRequested
+                    IsReadReceiptRequested     = $IsReadReceiptRequested
+                    FunctionName               = $MyInvocation.MyCommand
+                }
+
+                #region Put parameters (JSON Parts) into a valid "message"-JSON-object together
+                $bodyJsonParts = @()
+                $bodyHash = @{
+                    "message"         = (New-JsonMailObject @jsonParams)
+                    "saveToSentItems" = ($SaveToSentItems | ConvertTo-Json)
+                }
+                foreach ($key in $bodyHash.Keys) {
+                    $bodyJsonParts = $bodyJsonParts + """$($key)"" : $($bodyHash[$Key])"
+                }
+                $bodyJSON = "{`n" + ([string]::Join(",`n", $bodyJsonParts)) + "`n}"
+                #endregion Put parameters (JSON Parts) into a valid "message"-JSON-object together
+
+                if ($pscmdlet.ShouldProcess($Subject, "Send")) {
+                    Write-PSFMessage -Tag "MessageSend" -Level Verbose -Message "Send message with subject '$($Subject)' to recipient '$($ToRecipients)'"
+                    $invokeParam = @{
+                        "Field"        = "sendMail"
+                        "User"         = $User
+                        "Body"         = $bodyJSON
+                        "ContentType"  = "application/json"
+                        "Token"        = $Token
+                        "FunctionName" = $MyInvocation.MyCommand
+                    }
+                    $null = Invoke-MgaPostMethod @invokeParam
+                    if ($PassThru) { $messageItem.InputObject }
+                }
+            }
+
+            Default { stop-PSFMessage -Message "Unhandled parameter set. ($($PSCmdlet.ParameterSetName)) Developer mistake." -EnableException $true -Category "ParameterSetHandling" -FunctionName $MyInvocation.MyCommand }
         }
         #endregion send message
     }